+++ controllers/cart.controller.js
@@ -6,16 +6,29 @@ class CartController {
 
         const { productID, quantity } = req.body;
 
+        let result = {
+            message: null,
+            status: null,
+            data: null,
+        };
+
         if (!productID || !quantity) {
             res.status(400);
-            return res.json({ message: "Cart is empty" });
+            return res.json({ message: "Invalid input fields" });
         }
 
-        const result = await cartService.add(productID, quantity);
-        res.status(result.status);
+        try {
+            const data = await cartService.add(productID, quantity);
+            result.status = 200;
+            result.message = "Product added to cart!"
+            result.data = data;
+        } catch (error) {
+            result.status = 400; 
+            res.message = error.message;
+        } finally {
+            return res.json(result); 
+        }
 
-        // Return results from service
-        return res.json({ data: result.data, message: "Product added to cart!" });
     }
 }
