+++ routes/index.js
@@ -1,8 +1,11 @@
 const express = require('express');
-const publicRoutes = require('./public.routes'); 
 const app = express();
-
 app.use(express.json()); 
+
+//IMPORT ROUTES
+const publicRoutes = require('./public.routes'); 
+
+//CONFIGURE ROUTES
 app.use(publicRoutes);
 
 module.exports = app;
